Open in sidebar on shift click
If an entry in the calendar is clicked with shift pressed down,
open that day's journal in the sidebar.

To implement this, we have to keep track of whether the shift key is
pressed, so a document-level event handler was added.

Further, because the openInRightSidebar only accepts UUIDs,
we have to retrieve or create the day's page to get its UUID.

Resolves #4